revert accidental commit to configure.ac and fix error backporting NSDebug.m
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/branches/stable@28300 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSDebug.m
@@ -927,7 +927,7 @@ recover(int sig)
 }
 
 void *
-NSFrameAddress(NSUInteger offset)
+NSFrameAddress(int offset)
 {
   jbuf_type     *env;
 
@@ -987,7 +987,7 @@ NSFrameAddress(NSUInteger offset)
   return env->addr;
 }
 
-NSUInteger NSCountFrames(void)
+unsigned NSCountFrames(void)
 {
   jbuf_type	*env;
 
@@ -1051,7 +1051,7 @@ done:
 }
 
 void *
-NSReturnAddress(NSUInteger offset)
+NSReturnAddress(int offset)
 {
   jbuf_type	*env;
